refactor: feature:register를 feature:selectschool로 개명
