+++ feature/selectschool/src/main/AndroidManifest.xml
