+++ feature/selectschool/src/main/java/com/practice/selectschool/SelectSchoolUiState.kt
@@ -1,17 +1,17 @@
-package com.practice.register
+package com.practice.selectschool
 
 import com.practice.domain.School
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 
-data class RegisterUiState(
+data class SelectSchoolUiState(
     val selectedSchool: School,
     val schoolQuery: String,
     val schools: ImmutableList<School>,
 ) {
 
     companion object {
-        val Empty = RegisterUiState(
+        val Empty = SelectSchoolUiState(
             selectedSchool = School.EmptySchool,
             schoolQuery = "",
             schools = persistentListOf(),
