+++ settings.gradle.kts
@@ -23,7 +23,7 @@ include(":core:combine")
 include(":core:designsystem")
 include(":feature:main")
 include(":feature:onboarding")
-include(":feature:register")
+include(":feature:selectschool")
 include(":core:firebase")
 include(":core:util")
 include(":data:domain")
